[website_coupon] Factor out utility methods from controller to gift.coupon model
... to ease unit-testing and make the controller more readable and the logic reusable.

The coupon usage logic is also fixed: there was no clear rule to compute
the number of coupon uses from an order. It was an hard-coded 1 value
although the coupon amount was dependant of the coupon type (category),
but not of the product quantity. The rules are now clearly established
and unit-tested: a customer will consume as much coupons as available to
him to reduce its order as much as possible.