In base_accounting_kit/models/payment_matching.py (Odoo 14) there is no bank_account_id in the bank.statement.line  it is supposed to be partner_bank_id.
This bug is similar to the one documented here:  https://github.com/OCA/account-reconcile/pull/380

It is easily noticed / repeatable if you are trying to use camt.053 files which contain IBAN number of the partner involved in a payment.  The conditional will get hit, and you will always get the Error that bank_account_id is not part of the bank.statement.line model.  Which is indeed the case.  It should be partner_bank_id.